Fixes for CTE (de)serialization compatibility with older versions
