+++ src/include/duckdb/storage/serialization/query_node.json
@@ -21,6 +21,9 @@
         "name": "cte_map",
         "type": "CommonTableExpressionMap"
       }
+    ],
+    "finalize_deserialization": [
+        "ExtractCTENodes(result);"
     ]
   },
   {
@@ -176,13 +179,8 @@
         "id": 203,
         "name": "aliases",
         "type": "vector<string>"
-      },
-      {
-        "id": 204,
-        "name": "materialized",
-        "type": "CTEMaterialize",
-        "default": "CTEMaterialize::CTE_MATERIALIZE_DEFAULT"
       }
-    ]
+    ],
+    "custom_implementation": true
   }
 ]
